requirements.txt
holidays
dask[complete]>=2023.4.0
pandas>=1.5.0
# Numpy travado na 1.23.5 para não brigar com o TF 2.12
numpy==1.23.5
scikit-learn>=1.3.0
# Versões estáveis e leves para CPU
tensorflow-cpu==2.12.0
keras==2.12.0
pyyaml>=6.0
joblib>=1.3.0
pyarrow>=12.0.0
tqdm>=4.65.0